% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtp_isce.R
\name{CE.mixed.grad}
\alias{CE.mixed.grad}
\title{RTP p-value by cross-entropy importance sampling}
\usage{
CE.mixed.grad(p.val, ro = 0.01, J, theta = 1, N = 10^4, idx = 1)
}
\arguments{
\item{p.val}{Numeric vector of p-values.}

\item{ro}{Quantile level for cross-entropy updating.}

\item{J}{Number of smallest p-values used in the RTP statistic.}

\item{theta}{Initial proposal parameter.}

\item{N}{Number of ISCE sampling.}

\item{idx}{Decay parameter for mixture probabilities.}
}
\value{
Estimated RTP p-value.
}
\description{
Estimates an RTP p-value using cross-entropy importance sampling.
}
