% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtp_integral.R
\name{rtp_integral}
\alias{rtp_integral}
\title{RTP p-value by integral approximation}
\usage{
rtp_integral(p.values, i)
}
\arguments{
\item{p.values}{Numeric vector of p-values.}

\item{i}{Truncation point or vector of truncation points.}
}
\value{
Numeric RTP p-value or vector of RTP p-values.
}
\description{
Computes RTP p-values using an integral approximation.
}
